Migrate exp2 tests
