+++ src/webgpu/shader/execution/expression/call/builtin/exp2.spec.ts
@@ -11,10 +11,10 @@ import { makeTestGroup } from '../../../../../../common/framework/test_group.js'
 import { GPUTest } from '../../../../../gpu_test.js';
 import { kValue } from '../../../../../util/constants.js';
 import { TypeF32 } from '../../../../../util/conversion.js';
-import { exp2Interval } from '../../../../../util/f32_interval.js';
+import { FP } from '../../../../../util/floating_point.js';
 import { biasedRange, linearRange } from '../../../../../util/math.js';
 import { makeCaseCache } from '../../case_cache.js';
-import { allInputSources, generateUnaryToF32IntervalCases, run } from '../../expression.js';
+import { allInputSources, run } from '../../expression.js';
 
 import { builtin } from './builtin.js';
 
@@ -33,10 +33,10 @@ const inputs = [
 
 export const d = makeCaseCache('exp2', {
   f32_const: () => {
-    return generateUnaryToF32IntervalCases(inputs, 'finite', exp2Interval);
+    return FP.f32.generateScalarToIntervalCases(inputs, 'finite', FP.f32.exp2Interval);
   },
   f32_non_const: () => {
-    return generateUnaryToF32IntervalCases(inputs, 'unfiltered', exp2Interval);
+    return FP.f32.generateScalarToIntervalCases(inputs, 'unfiltered', FP.f32.exp2Interval);
   },
 });
